# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Attendance Gateway API",
    "version": "1.4.0",
    "description": "Zero-Trust Cryptographic Attendance Gateway.\n\n**Authentication**\n- Most client endpoints accept a shared client API key: header `X-Api-Key: ag_...` (mint from the Admin console → API Keys).\n- Role-protected endpoints (professor/admin) use a JWT: `Authorization: Bearer <token>` from `/api/v1/admin/login` or `/api/v1/auth/login`.\n- Global security: `ApiKey OR bearer`; public operations (health, admin login) override with `[]`.\n\n**Building a client?** See the **Client Integration Guide** (QR / 3-second metronome protocol, biometric Gate 2 ordering, timestamp sealing, 250ms window): https://git.atmyhome.tech/het/attendance-gateway/src/branch/main/docs/CLIENT_INTEGRATION_GUIDE.md\n\n**Error payloads:** every 4xx/5xx returns the envelope `{success:false,error:{code,message,latency_ms?,details?}}`. Full code table + client UI actions: **Error Dictionary** — https://git.atmyhome.tech/het/attendance-gateway/src/branch/main/docs/ERROR_DICTIONARY.md (schema: `ErrorEnvelope`).\n\n**To get an API key:** sign in to the Admin console (https://admin.atmyhome.tech) → **API Keys** → Generate Key."
  },
  "servers": [
    {
      "url": "https://api.atmyhome.tech"
    }
  ],
...
  "security": [
    {
      "ApiKey": []
    },
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Shared client API key (mint from the Admin console)."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "ERR_HW_MISMATCH",
                  "ERR_SIG_INVALID",
                  "ERR_STREAM_DETECTED",
                  "ERR_TOKEN_EXPIRED",
                  "ERR_NONCE_USED",
                  "ERR_NONCE_INVALID",
                  "ERR_AUTH_MISSING",
                  "ERR_FORBIDDEN",
                  "ERR_NOT_FOUND",
                  "ERR_RATE_LIMIT",
                  "ERR_BAD_REQUEST",
                  "ERR_CONFLICT",
                  "ERR_INTERNAL",
                  "ERR_UNAVAILABLE",
                  "ERR_UNKNOWN"
                ]
              },
              "message": {
                "type": "string"
              },
              "latency_ms": {
                "type": "number",
                "description": "Present only for ERR_STREAM_DETECTED."
              },
              "details": {
                "type": "object"
              }
            }
          }
        },
        "example": {
          "success": false,
          "error": {
            "code": "ERR_STREAM_DETECTED",
            "message": "Stream artifact detected",
            "latency_ms": 310
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/claim-attendance": {
      "post": {
        "tags": [
          "Attendance / Claim"
        ],
        "summary": "Submit an attendance claim (Gate 1-4 pipeline; HMAC-signed, ≤250ms TTL enforced)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "session_uuid": "uuid",
                "student_uuid": "uuid",
                "token_val": "Ab3d",
                "client_claimed_time": 1720000000123,
                "device_id_hash": "sha256-hex",
                "nonce": "hex",
                "signature": "hmac-hex"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PRESENT",
            "content": {
              "application/json": {
                "example": {
                  "status": "PRESENT",
                  "verification_delta_ms": 87,
                  "ledger_uuid": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Anti-replay: nonce consumed (Gate 4)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "ERR_NONCE_USED",
                    "message": "Anti-replay protection triggered. Nonce already consumed."
                  }
                }
              }
            }
          },
          "401": {
            "description": "Forged/corrupted HMAC signature (Gate 4)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "ERR_SIG_INVALID",
                    "message": "Cryptographic rejection. The HMAC signature was forged or corrupted."
                  }
                }
              }
            }
          },
          "403": {
            "description": "Hardware mismatch (Gate 1)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "ERR_HW_MISMATCH",
                    "message": "Unregistered hardware. Your account is bound to a different physical device."
                  }
                }
              }
            }
          },
          "406": {
            "description": "Visual token expired/invalid (Gate 3)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "ERR_TOKEN_EXPIRED",
                    "message": "The scanned visual token is expired or invalid."
                  }
                }
              }
            }
          },
          "412": {
            "description": "Proxy stream detected / time drift (Gate 4)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "ERR_STREAM_DETECTED",
                    "message": "Proxy Stream Detected. Photonic intercept lag exceeded the 250ms threshold.",
                    "latency_ms": 310
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sessions": {
      "get": {
        "tags": [
          "Sessions / Professor"
        ],
        "summary": "List sessions for professor (JWT-locked)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "sessions"
          }
        }
      }
    },
    "/api/v1/sessions/start": {
      "post": {
        "tags": [
          "Sessions / Professor"
        ],
        "summary": "Start a session (mints live 4-char token metronome)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "course_code": "CS101"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "session started"
          }
        }
      }
    },
    "/api/v1/sessions/{sessionUuid}": {
      "get": {
        "tags": [
          "Sessions / Professor"
        ],
        "summary": "Session detail",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "session"
          }
        }
      }
    },
    "/api/v1/sessions/{sessionUuid}/stop": {
      "post": {
        "tags": [
          "Sessions / Professor"
        ],
        "summary": "Stop a session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "stopped"
          }
        }
      }
    },
    "/api/v1/sessions/{sessionUuid}/tokens": {
      "get": {
        "tags": [
          "Sessions / Professor"
        ],
        "summary": "Current token + schedule (APK fallback poll)",
        "parameters": [
          {
            "name": "sessionUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "token"
          }
        }
      }
    },
    "/api/v1/sessions/{sessionUuid}/challenge": {
      "get": {
        "tags": [
          "Attendance / Claim"
        ],
        "summary": "Issues a server nonce for the claim (anti-replay)",
        "parameters": [
          {
            "name": "sessionUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "nonce"
          }
        }
      }
    },
    "/api/v1/sessions/{sessionUuid}/attendance": {
      "get": {
        "tags": [
          "Sessions / Professor"
        ],
        "summary": "Roster attendance for a session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "roster"
          }
        }
      }
    },
    "/api/v1/professor/timetable": {
      "get": {
        "tags": [
          "Sessions / Professor"
        ],
        "summary": "Today's timetable for logged-in professor",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "timetable"
          }
        }
      }
    },
    "/api/v1/professor/summary": {
      "get": {
        "tags": [
          "Sessions / Professor"
        ],
        "summary": "Attendance summary for logged-in professor",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "summary"
          }
        }
      }
    },
    "/api/v1/students/{studentUuid}/attendance": {
      "get": {
        "tags": [
          "Sessions / Professor"
        ],
        "summary": "A student's attendance ledger",
        "parameters": [
          {
            "name": "studentUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ledger"
          }
        }
      }
    },
...
  }
}